// web-client/iron-svelte-replay-client/src/types/recording.types.d.ts
type milliseconds = number;
export interface Header {
    version: number;
    duration: milliseconds;
    totalPdus: number;
}
export interface IndexTableRow {
    timeOffset: milliseconds;
    pduLength: number;
    byteOffset: bigint;
    direction: number;
}
export interface PDUEntry {
    data: ArrayBuffer;
    timeOffset: milliseconds;
}
export {};